requirements.txt
asgiref==3.7.2
boto3==1.27.0
botocore==1.30.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.1.0
django-dotenv==1.4.2
django-environ==0.10.0
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
future==0.18.3
geocoder==1.38.1
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
ratelim==0.1.6
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
